Testa excecoes gerados em uma aplicacao python
