+++ testes-com-pytest/funcoes_numeros.py
@@ -1,6 +1,7 @@
 # checa se um número é impar ou par
 def eh_impar(numero):
-  return numero % 2 != 0
+    return numero % 2 != 0
+
 
 def dividir(dividendo, divisor):
-  return dividendo / divisor
+    return dividendo / divisor
